==> nbs/05_faq_prepare_markdown.ipynb <==
#| default_exp scripts/faq_prepare_markdown

#| export
import argparse
import sys
import pandas as pd

#| export
def process(file_name_faq: str, file_name_faq_markdown: str):
    df = pd.read_csv(file_name_faq)
    df["content"] = "Question: " + df["question"] + "\n\n" + "Answer: " + df["answer"]
    df["url"] = "FAQ"
    df[["content", "url"]].to_json(file_name_faq_markdown, orient="records", lines=True)

process(
    "../data/faq.csv",
    "../data/faq--markdown.jsonl",
)

# | export
if __name__ == "__main__" and "ipykernel_launcher" not in " ".join(sys.argv) and "nbdev" not in " ".join(sys.argv):
    parser = argparse.ArgumentParser()
    parser.add_argument("--file_name_faq",
                        type=str,
                        required=True,
                        help="CSV with FAQ")
    parser.add_argument("--file_name_faq_markdown",
                        type=str,
                        required=True,
                        help="JSONL file with FAQ as markdown")
    process(**vars(parser.parse_args()))

#| hide
import nbdev; nbdev.nbdev_export()
