==> nbs/06_concatenate_openai_files_lists.ipynb <==
#| default_exp scripts/concatenate_openai_files_list

#| export
import argparse
import sys
from typing import List
import pandas as pd

#| export
MAX_FILE_PER_ASSISTENT = 20

#| export
def process(file_names_input: List[str], file_name_output: str):
    df = pd.concat([
        pd.read_json(fname, orient="records", lines=True)
        for fname in file_names_input
    ])
    df = df.reset_index(drop=True)
    assert len(df) < MAX_FILE_PER_ASSISTENT
    df.to_json(file_name_output, orient="records", lines=True)

process(
    [
        "../data/urls--downloaded--openai-file-ids.jsonl",
        "../data/faq--openai-file-ids.jsonl",
    ],
    "../data/openai-file-ids.jsonl"
)

# | export
if __name__ == "__main__" and "ipykernel_launcher" not in " ".join(sys.argv) and "nbdev" not in " ".join(sys.argv):
    parser = argparse.ArgumentParser()
    parser.add_argument("--file_names_input",
                        nargs="+",
                        required=True,
                        help="A list with OpenAI files ids")
    parser.add_argument("--file_name_output",
                        type=str,
                        required=True,
                        help="JSONL file with all the file ids")
    process(**vars(parser.parse_args()))

#| hide
import nbdev; nbdev.nbdev_export()
